Refactor ExecutionFunctions
