+++ client/src/route/digitaltwins/pipelineHandler.ts
@@ -1,0 +1,132 @@
+import { Dispatch, SetStateAction } from 'react';
+import { AlertColor } from '@mui/material';
+import DigitalTwin, { formatName } from 'util/gitlabDigitalTwin';
+import { useDispatch } from 'react-redux';
+import {
+  startPipeline,
+  updatePipelineState,
+  updatePipelineStateOnStop,
+} from './pipelineUtils';
+import { checkFirstPipelineStatus } from './pipelineChecks';
+
+export const handleButtonClick = (
+  buttonText: string,
+  setButtonText: Dispatch<SetStateAction<string>>,
+  digitalTwin: DigitalTwin,
+  setSnackbarMessage: Dispatch<SetStateAction<string>>,
+  setSnackbarSeverity: Dispatch<SetStateAction<AlertColor>>,
+  setSnackbarOpen: Dispatch<SetStateAction<boolean>>,
+  setLogButtonDisabled: Dispatch<SetStateAction<boolean>>,
+  dispatch: ReturnType<typeof useDispatch>,
+) => {
+  if (buttonText === 'Start') {
+    handleStart(
+      buttonText,
+      setButtonText,
+      setSnackbarMessage,
+      setSnackbarSeverity,
+      setSnackbarOpen,
+      digitalTwin,
+      setLogButtonDisabled,
+      dispatch,
+    );
+  } else {
+    handleStop(
+      digitalTwin,
+      setSnackbarMessage,
+      setSnackbarSeverity,
+      setSnackbarOpen,
+      setButtonText,
+      dispatch,
+    );
+  }
+};
+
+export const handleStart = async (
+  buttonText: string,
+  setButtonText: Dispatch<SetStateAction<string>>,
+  setSnackbarMessage: Dispatch<SetStateAction<string>>,
+  setSnackbarSeverity: Dispatch<SetStateAction<AlertColor>>,
+  setSnackbarOpen: Dispatch<SetStateAction<boolean>>,
+  digitalTwin: DigitalTwin,
+  setLogButtonDisabled: Dispatch<SetStateAction<boolean>>,
+  dispatch: ReturnType<typeof useDispatch>,
+) => {
+  if (buttonText === 'Start') {
+    setButtonText('Stop');
+    setLogButtonDisabled(true);
+    updatePipelineState(digitalTwin, dispatch);
+    await startPipeline(
+      digitalTwin,
+      setSnackbarMessage,
+      setSnackbarSeverity,
+      setSnackbarOpen,
+      setLogButtonDisabled,
+      dispatch,
+    );
+    const params = {
+      setButtonText,
+      digitalTwin,
+      setLogButtonDisabled,
+      dispatch,
+    };
+    checkFirstPipelineStatus(params);
+  } else {
+    setButtonText('Start');
+  }
+};
+
+export const handleStop = async (
+  digitalTwin: DigitalTwin,
+  setSnackbarMessage: Dispatch<SetStateAction<string>>,
+  setSnackbarSeverity: Dispatch<SetStateAction<AlertColor>>,
+  setSnackbarOpen: Dispatch<SetStateAction<boolean>>,
+  setButtonText: Dispatch<SetStateAction<string>>,
+  dispatch: ReturnType<typeof useDispatch>,
+) => {
+  try {
+    await stopPipelines(digitalTwin);
+    setSnackbar(
+      `Execution stopped successfully for ${formatName(digitalTwin.DTName)} (Run #${digitalTwin.executionCount})`,
+      'success',
+      setSnackbarMessage,
+      setSnackbarSeverity,
+      setSnackbarOpen,
+    );
+  } catch (error) {
+    setSnackbar(
+      `Execution stop failed for ${digitalTwin.DTName} (Run #${digitalTwin.executionCount})`,
+      'error',
+      setSnackbarMessage,
+      setSnackbarSeverity,
+      setSnackbarOpen,
+    );
+  } finally {
+    updatePipelineStateOnStop(digitalTwin, setButtonText, dispatch);
+  }
+};
+
+const stopPipelines = async (digitalTwin: DigitalTwin) => {
+  if (digitalTwin.gitlabInstance.projectId && digitalTwin.pipelineId) {
+    await digitalTwin.stop(
+      digitalTwin.gitlabInstance.projectId,
+      digitalTwin.pipelineId,
+    );
+    await digitalTwin.stop(
+      digitalTwin.gitlabInstance.projectId,
+      digitalTwin.pipelineId + 1,
+    );
+  }
+};
+
+const setSnackbar = (
+  message: string,
+  severity: AlertColor,
+  setSnackbarMessage: Dispatch<SetStateAction<string>>,
+  setSnackbarSeverity: Dispatch<SetStateAction<AlertColor>>,
+  setSnackbarOpen: Dispatch<SetStateAction<boolean>>,
+) => {
+  setSnackbarMessage(message);
+  setSnackbarSeverity(severity);
+  setSnackbarOpen(true);
+};
